set to HI timezone

+++ modules/process_data.py
@@ -29,14 +29,15 @@ def update_arrivals_db(arrival_dict):
 	ssl_key_path = "client-key.pem"
 	ssl_root_cert_path = "server-ca.pem"
 
-	with open(ssl_cert_path, 'w+') as f:
-		f.write(os.environ["SSL_CERT"])
+	if not os.path.exists(ssl_cert_path):
+		with open(ssl_cert_path, 'w+') as f:
+			f.write(os.environ["SSL_CERT"])
 
-	with open(ssl_key_path, 'w+') as f:
-		f.write(os.environ["SSL_KEY"])
+		with open(ssl_key_path, 'w+') as f:
+			f.write(os.environ["SSL_KEY"])
 
-	with open(ssl_root_cert_path, 'w+') as f:
-		f.write(os.environ["SSL_ROOT_CERT"])
+		with open(ssl_root_cert_path, 'w+') as f:
+			f.write(os.environ["SSL_ROOT_CERT"])
 
 	conn = psycopg2.connect(database=str(os.environ["DB_NAME"]),
 							user=str(os.environ["DB_USERNAME"]),
